requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pillow==10.4.0
numpy==1.24.4
python-multipart
streamlit
requests

# 🔥 ADD THIS LINE (VERY IMPORTANT)
blinker==1.4